requirements.txt
opencv-python
deepface
numpy
mediapipe
pandas
tensorflow
psycopg2-binary
pgvector
